[Fix] campDetail에 이미지가 null값으로 가져와지던 오류 해결
api연동후 이미지가 null값으로 가져와지던 오류 해결.
