+++ app/src/main/java/com/brandon/campingmate/network/retrofit/CampImage.kt
@@ -27,7 +27,7 @@ data class CampImageItems(
 data class CampItem(
     val contentId: Int?,
     val serialnum: Int?,
-    val imageurl: String?,
+    val imageUrl: String?,
     val createdtime: String?,
     val modifiedtime: String?,
 )
