+++ app/src/main/res/layout/fragment_search.xml
@@ -151,6 +151,7 @@
                     android:id="@+id/chip_type_group"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
+                    app:singleSelection="true"
                     app:layout_constraintStart_toStartOf="@id/tv_search_type"
                     app:layout_constraintTop_toBottomOf="@+id/tv_search_type">
 
@@ -676,6 +677,7 @@
                     android:id="@+id/chip_thema_group"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
+                    app:singleSelection="true"
                     app:layout_constraintStart_toStartOf="@id/tv_search_thema"
                     app:layout_constraintTop_toBottomOf="@+id/tv_search_thema">
 
@@ -830,6 +832,7 @@
                     android:id="@+id/chip_bottom_group"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
+                    app:singleSelection="true"
                     app:layout_constraintStart_toStartOf="@id/tv_search_bottom"
                     app:layout_constraintTop_toBottomOf="@+id/tv_search_bottom">
